+++ changelog.md
@@ -1,8 +1,9 @@
 # 1.3.0
 - April Fools is over :(
+- Add __Kingsammelot's__ completion of [Killbot](https://youtu.be/I1_yfoHOrbQ)
+- Add *Custom Reactions*
 - Add option to disable the reactions on platformer
 - Make end trigger support better
-- Add __Kingsammelot's__ completion of [Killbot](https://youtu.be/I1_yfoHOrbQ)
 
 # 1.2.1
 - wat